+++ model/src/test/scala/org/broadinstitute/dsde/workbench/model/ErrorReportSpec.scala
@@ -7,8 +7,9 @@ import spray.json._
 import ErrorReportJsonSupport._
 
 class ErrorReportSpec extends FlatSpecLike with BeforeAndAfterAll with Matchers {
-  "ErrorReport" should "roundtrip json" in {
-    implicit val errorReportSource = ErrorReportSource("test")
+  implicit val errorReportSource = ErrorReportSource("test")
+
+  "json serialization" should "roundtrip" in {
     val internalEr = ErrorReport("internalERMessage")
     val stacktrace = Seq(new StackTraceElement("testClass", "testMethod", "fileName", 42))
     val er = ErrorReport("testMessage", Some(StatusCodes.OK), Seq(internalEr), stacktrace, Some(classOf[RuntimeException]))
@@ -17,5 +18,44 @@ class ErrorReportSpec extends FlatSpecLike with BeforeAndAfterAll with Matchers 
     val erRead = erJson.convertTo[ErrorReport]
 
     erRead shouldBe er
+  }
+
+  "message function" should "return exception message if it exists" in {
+    val msgExc = new RuntimeException("boom")
+    ErrorReport.message(msgExc) shouldBe "boom"
+  }
+
+  it should "return exception class if no message exists" in {
+    val noMsgExc = new RuntimeException()
+    ErrorReport.message(noMsgExc) shouldBe "RuntimeException"
+  }
+
+  "causes function" should "return an ErrorReport with no causes if there are none" in {
+    val exc = new RuntimeException("boom")
+    ErrorReport.causes(exc) shouldBe empty
+  }
+
+  it should "return an ErrorReport with suppressed exceptions if there are any" in {
+    val excSuppressed = new RuntimeException("suppressed")
+    val exc = new RuntimeException("boom")
+    exc.addSuppressed(excSuppressed)
+
+    ErrorReport.causes(exc) shouldBe Seq(ErrorReport(excSuppressed))
+  }
+
+  it should "return an ErrorReport with suppressed exceptions even if there are other causes" in {
+    val excSuppressed = new RuntimeException("suppressed")
+    val excCause = new RuntimeException("cause")
+    val exc = new RuntimeException("boom", excCause)
+    exc.addSuppressed(excSuppressed)
+
+    ErrorReport.causes(exc) shouldBe Seq(ErrorReport(excSuppressed))
+  }
+
+  it should "return an ErrorReport with causes if there are no suppressed exceptions" in {
+    val excCause = new RuntimeException("cause")
+    val exc = new RuntimeException("boom", excCause)
+
+    ErrorReport.causes(exc) shouldBe Seq(ErrorReport(excCause))
   }
 }