+++ scripts/publish.sh
@@ -2,8 +2,11 @@
 
 set -e
 
+# sbt publish publishes libs to Artifactory for the scala version sbt is running as.
+# sbt +publish publishes libs to Artifactory for all scala versions listed in crossScalaVersions.
+# We only do sbt publish here because Travis runs against 2.11 and 2.12 in separate jobs, so each one publishes its version to Artifactory.
 if [[ "$TRAVIS_PULL_REQUEST" == "false" && "$TRAVIS_BRANCH" == "develop" ]]; then
-	sbt +publish -Dproject.isSnapshot=false
+	sbt publish -Dproject.isSnapshot=false
 else
-	sbt +publish -Dproject.isSnapshot=true
+	sbt publish -Dproject.isSnapshot=true
 fi