+++ util/src/main/scala/org/broadinstitute/dsde/workbench/util/FutureSupport.scala
@@ -2,7 +2,7 @@ package org.broadinstitute.dsde.workbench.util
 
 import java.util.concurrent.TimeoutException
 
-import akka.actor.ActorContext
+import akka.actor.{ActorContext, Scheduler}
 import akka.pattern.after
 
 import scala.concurrent.duration.FiniteDuration
@@ -28,7 +28,11 @@ trait FutureSupport {
    */
   def assertSuccessfulTries[K, T](tries: Map[K, Try[T]])(implicit executionContext: ExecutionContext): Future[Map[K, T]] = {
     val failures = tries.values.collect{ case Failure(t) => t }
-    if (failures.isEmpty) Future.successful(tries.map { case (k, v) => k -> v.get}) else Future.failed(failures.head)
+    if (failures.isEmpty) {
+      Future.successful(tries.mapValues(_.get))
+    } else { 
+      Future.failed(failures.head)
+    }
   }
 
   /**
@@ -41,8 +45,8 @@ trait FutureSupport {
     * }}}
     */
   implicit class FutureWithTimeout[A](f: Future[A]) {
-    def withTimeout(duration: FiniteDuration, errMsg: String)(implicit context: ActorContext, ec: ExecutionContext): Future[A] =
-      Future.firstCompletedOf(Seq(f, after(duration, context.system.scheduler)(Future.failed(new TimeoutException(errMsg)))))
+    def withTimeout(duration: FiniteDuration, errMsg: String)(implicit scheduler: Scheduler, ec: ExecutionContext): Future[A] =
+      Future.firstCompletedOf(Seq(f, after(duration, scheduler)(Future.failed(new TimeoutException(errMsg)))))
   }
 }