+++ util/src/main/scala/org/broadinstitute/dsde/workbench/util/package.scala
@@ -8,15 +8,13 @@ import scala.concurrent.duration._
 package object util {
   def toScalaDuration(javaDuration: java.time.Duration) = Duration.fromNanos(javaDuration.toNanos)
 
-  def addJitter(baseTime: FiniteDuration, maxJitterToAdd: FiniteDuration): FiniteDuration = {
+  def addJitter(baseTime: FiniteDuration, maxJitterToAdd: Duration): FiniteDuration = {
     baseTime + ((scala.util.Random.nextFloat * maxJitterToAdd.toNanos) nanoseconds)
   }
 
   def addJitter(baseTime: FiniteDuration): FiniteDuration = {
-    if(baseTime < (1 second)) {
-      addJitter(baseTime, 100 milliseconds)
-    } else if (baseTime < (10 seconds)) {
-      addJitter(baseTime, 500 milliseconds)
+    if(baseTime <= (10 seconds)) {
+      addJitter(baseTime, baseTime * 0.1)
     } else {
       addJitter(baseTime, 1 second)
     }